Address OSMO experiment review feedback
Use the shared Arena image and scheduler defaults, and keep the submission API with its executable entry point.
